requirements.txt
discord-py-interactions==4.3.4
mcstatus==10.0.1
python-dotenv==0.21.0
rcon==2.3.6
